Revert "Merge branch 'master' of https://github.com/farshadfmr/opencart"
This reverts commit ae44126fce20cb29531c96edd774e2e91b542d3f.

+++ upload/catalog/model/checkout/cart.php
@@ -93,11 +93,11 @@ class Cart extends \Opencart\System\Engine\Model {
 	 *
 	 * @param array<int, array<string, mixed>> $totals
 	 * @param array<int, float>                $taxes
-	 * @param float                            $total
+	 * @param int                              $total
 	 *
 	 * @return void
 	 */
-	public function getTotals(array &$totals, array &$taxes, float &$total): void {
+	public function getTotals(array &$totals, array &$taxes, int &$total): void {
 		$sort_order = [];
 
 		// Extensions
